+++ backend/api/metrics/handler.go
@@ -2,4 +2,4 @@ package metrics
 
 type MetricsHandler struct{}
 
-var VERSION = "0.0.263"
+var VERSION = "0.0.264"
